Add chatroomId to popularSearch
