+++ src/main/java/dec/haeyum/calendar/service/impl/CalendarServiceImpl.java
@@ -10,6 +10,7 @@ import dec.haeyum.calendar.entity.CalendarEntity;
 import dec.haeyum.calendar.repository.CalendarRepository;
 import dec.haeyum.calendar.service.CalendarService;
 import dec.haeyum.chat.service.ChatService;
+import dec.haeyum.chat.service.ChatroomService;
 import dec.haeyum.config.error.ErrorCode;
 import dec.haeyum.config.error.exception.BusinessException;
 import dec.haeyum.movie.entity.MovieEntity;
@@ -39,7 +40,7 @@ public class CalendarServiceImpl implements CalendarService {
 
     private final CalendarRepository calendarRepository;
     private final SongService songService;
-    private final ChatService chatService;
+    private final ChatroomService chatroomService;
 
     @Override
     @Transactional
@@ -85,7 +86,7 @@ public class CalendarServiceImpl implements CalendarService {
 
         List<ResponseCalendarDto> responseItems = new ArrayList<>();
         for (CalendarEntity calendar : paging.getContent()) {
-            Long chatroomId = chatService.getChatRoomIdByCalendar(calendar);
+            Long chatroomId = chatroomService.getChatRoomIdByCalendar(calendar);
             String calendarSongImageUrl = songService.getCalendarSongImageUrl(calendar.getCalendarId());
             ResponseCalendarDto responseCalendarDto = new ResponseCalendarDto(calendar, calendarSongImageUrl, chatroomId);
             responseItems.add(responseCalendarDto);
