+++ src/main/java/dec/haeyum/chat/service/ChatService.java
@@ -14,5 +14,5 @@ public interface ChatService {
 
     void createChatRoom();
 
-    Long getChatRoomIdByCalendar(CalendarEntity calendar);
+
 }
