+++ src/main/java/dec/haeyum/chat/service/ChatroomService.java
@@ -1,0 +1,7 @@
+package dec.haeyum.chat.service;
+
+import dec.haeyum.calendar.entity.CalendarEntity;
+
+public interface ChatroomService {
+    Long getChatRoomIdByCalendar(CalendarEntity calendar);
+}
